market_scraper/requirements.txt
yfinance>=0.2.0
pymongo>=4.6.0
python-dotenv>=1.0.0

